@@ -1,25 +1,25 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <UserSecretsId>bw</UserSecretsId>
-    <LangVersion>latest</LangVersion>
+    <LangVersion>preview</LangVersion>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App"/>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1"/>
-    <PackageReference Include="BioEngine.Core.Infra" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Core.Infra')"/>
-    <PackageReference Include="BioEngine.Core.API" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Core.API')"/>
-    <PackageReference Include="BioEngine.BRC.Domain" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.BRC.Domain')"/>
-    <PackageReference Include="BioEngine.Extra.IPB" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Extra.IPB')"/>
-    <PackageReference Include="BioEngine.Extra.Twitter" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Extra.Twitter')"/>
-    <PackageReference Include="BioEngine.Extra.Facebook" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Extra.Facebook')"/>
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc2" />
+    <PackageReference Include="BioEngine.Core.Infra" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Core.Infra')" />
+    <PackageReference Include="BioEngine.Core.API" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Core.API')" />
+    <PackageReference Include="BioEngine.BRC.Domain" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.BRC.Domain')" />
+    <PackageReference Include="BioEngine.Extra.IPB" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Extra.IPB')" />
+    <PackageReference Include="BioEngine.Extra.Twitter" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Extra.Twitter')" />
+    <PackageReference Include="BioEngine.Extra.Facebook" Version="3.0.*" Condition="!Exists('..\..\..\BioEngine.Extra.Facebook')" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\..\BioEngine.Core.Infra\src\BioEngine.Core.Infra\BioEngine.Core.Infra.csproj" Condition="Exists('..\..\..\BioEngine.Core.Infra')"/>
-    <ProjectReference Include="..\..\..\BioEngine.Core.API\src\BioEngine.Core.API\BioEngine.Core.API.csproj" Condition="Exists('..\..\..\BioEngine.Core.API')"/>
-    <ProjectReference Include="..\..\..\BioEngine.BRC.Domain\src\BioEngine.BRC.Domain\BioEngine.BRC.Domain.csproj" Condition="Exists('..\..\..\BioEngine.BRC.Domain')"/>
-    <ProjectReference Include="..\..\..\BioEngine.Extra.IPB\src\BioEngine.Extra.IPB\BioEngine.Extra.IPB.csproj" Condition="Exists('..\..\..\BioEngine.Extra.IPB')"/>
-    <ProjectReference Include="..\..\..\BioEngine.Extra.Twitter\src\BioEngine.Extra.Twitter\BioEngine.Extra.Twitter.csproj" Condition="Exists('..\..\..\BioEngine.Extra.Twitter')"/>
-    <ProjectReference Include="..\..\..\BioEngine.Extra.Facebook\src\BioEngine.Extra.Facebook\BioEngine.Extra.Facebook.csproj" Condition="Exists('..\..\..\BioEngine.Extra.Facebook')"/>
+    <ProjectReference Include="..\..\..\BioEngine.Core.Infra\src\BioEngine.Core.Infra\BioEngine.Core.Infra.csproj" Condition="Exists('..\..\..\BioEngine.Core.Infra')" />
+    <ProjectReference Include="..\..\..\BioEngine.Core.API\src\BioEngine.Core.API\BioEngine.Core.API.csproj" Condition="Exists('..\..\..\BioEngine.Core.API')" />
+    <ProjectReference Include="..\..\..\BioEngine.BRC.Domain\src\BioEngine.BRC.Domain\BioEngine.BRC.Domain.csproj" Condition="Exists('..\..\..\BioEngine.BRC.Domain')" />
+    <ProjectReference Include="..\..\..\BioEngine.Extra.IPB\src\BioEngine.Extra.IPB\BioEngine.Extra.IPB.csproj" Condition="Exists('..\..\..\BioEngine.Extra.IPB')" />
+    <ProjectReference Include="..\..\..\BioEngine.Extra.Twitter\src\BioEngine.Extra.Twitter\BioEngine.Extra.Twitter.csproj" Condition="Exists('..\..\..\BioEngine.Extra.Twitter')" />
+    <ProjectReference Include="..\..\..\BioEngine.Extra.Facebook\src\BioEngine.Extra.Facebook\BioEngine.Extra.Facebook.csproj" Condition="Exists('..\..\..\BioEngine.Extra.Facebook')" />
   </ItemGroup>
 </Project>
